#62 code is now generic, next step is to update the config file accordingly

+++ R/palmsplusr_shiny.R
@@ -6,6 +6,7 @@
 #' @param outputdir Path to outputdir location
 #' @param dataset_name Name of dataset
 #' @param configfile Configuration file
+#' @param verbose verbose Boolean
 #' @return palms_to_clean_lower object
 #' @importFrom stats end start formula as.formula
 #' @importFrom tidyr pivot_wider
@@ -21,7 +22,8 @@ palmsplusr_shiny <- function(gisdir = "",
                              gislinkfile = "",
                              outputdir = "",
                              dataset_name = "",
-                             configfile = "") {
+                             configfile = "",
+                             verbose = TRUE) {
   
   #-----------------------------
   # NEW CODE
@@ -70,7 +72,7 @@ palmsplusr_shiny <- function(gisdir = "",
   
   palmsplus_folder = paste0(outputdir, "/palmsplusr_output")
   if (!dir.exists(palmsplus_folder)) {
-    cat("\nCreating PALMSplusR output directory\n")
+    if (verbose) cat("\nCreating PALMSplusR output directory\n")
     dir.create(palmsplus_folder)
   }
   sf::sf_use_s2(FALSE)
@@ -91,14 +93,14 @@ palmsplusr_shiny <- function(gisdir = "",
   ))
   PALMS_combined <- bind_rows(csv_palms)
   # Data cleaning:
-  cat("\nstart cleaning\n")
+  if (verbose) cat("\nstart cleaning\n")
   PALMS_reduced <- subset(PALMS_combined, lon > -180)
   palms_reduced_cleaned <- check_and_clean_palms_data(PALMS_reduced, dataset_name)
-  cat("\ncleaning completed\n")
+  if (verbose) cat("\ncleaning completed\n")
   
   # Write to csv and read using read_palms to format the object as expected from the rest of the code
   PALMS_reduced_file = paste0(palmsplus_folder, "/", stringr::str_interp("PALMS_${dataset_name}_reduced.csv"))
-  cat(paste0("\nCheck PALMS_reduced_file: ", PALMS_reduced_file))
+  if (verbose) cat(paste0("\nCheck PALMS_reduced_file: ", PALMS_reduced_file))
   write.csv(palms_reduced_cleaned, PALMS_reduced_file)
   palms = palmsplusr::read_palms(PALMS_reduced_file)
   
@@ -108,7 +110,7 @@ palmsplusr_shiny <- function(gisdir = "",
     file_of_interest = allcsvfiles[which(tolower(basename(allcsvfiles)) == namelowercase)]
     return(file_of_interest)
   }
-  cat("\nreading basis file\n")
+  if (verbose) cat("\nreading basis file\n")
   participant_basis = read_csv(gislinkfile)
   unique_ids_in_palms <- unique(palms$identifier)
   unique_ids_in_participant_basis <- unique(participant_basis$identifier)
@@ -122,7 +124,6 @@ palmsplusr_shiny <- function(gisdir = "",
     loca[[jj]][[1]] = sf::read_sf(loca[[jj]][3]) #home_nbh
     loca[[jj]][[2]] = sf::read_sf(loca[[jj]][4]) #school_nbh
   }
-  
   
   # Force id numbers to be characrer(
   locationNames = names(loca)
@@ -153,9 +154,11 @@ palmsplusr_shiny <- function(gisdir = "",
   
   # Check for missing IDs -------------------------------------------------------------------------
   withoutMissingId = hbt_check_missing_id(participant_basis, palmsplus_folder, dataset_name, palms,
-                                          loca, groupinglocation = groupinglocation)
+                                          loca, groupinglocation = groupinglocation,
+                                          verbose = verbose)
   palms = withoutMissingId$palms
   participant_basis = withoutMissingId$participant_basis
+  loca = withoutMissingId$loca
   
   
 
@@ -174,7 +177,6 @@ palmsplusr_shiny <- function(gisdir = "",
   # Note: I have removed the dependency on palmsplusr for this as it
   # involved super assignment operators which seem to be causing issues,
   # defaults are now taken care of in the config file preparation
-  
   
   # #=============================
   # adding fields
@@ -215,11 +217,13 @@ palmsplusr_shiny <- function(gisdir = "",
     if (file.exists(fn)) file.remove(fn)
   }
   
+  
+
   Nlocation_objects = NULL
   for (i in 1:Nlocations) {
     Nlocation_objects = c(Nlocation_objects, length(loca[[i]][[1]]), length(loca[[i]][[2]]))
   }
-  cat("\n<<< building palmsplus...\n")
+  if (verbose) cat("\n<<< building palmsplus...\n")
   if (length(palms) > 0 & length(palmsplus_fields) &
       all(Nlocation_objects > 0) & length(participant_basis) > 0) {
     
@@ -227,63 +231,79 @@ palmsplusr_shiny <- function(gisdir = "",
     palmsplus <- hbt_build_palmsplus(data = palms, 
                                      palmsplus_fields = palmsplus_fields,
                                      loca = loca,
-                                     participant_basis = participant_basis)
+                                     participant_basis = participant_basis,
+                                     verbose = verbose)
     write_csv(palmsplus, file = fns[1])
-    cat("done>>>\n")
+    if (verbose) cat(">>>\n")
   } else {
-    cat("skipped because insufficient input data>>>\n")
+    if (verbose) cat("skipped because insufficient input data>>>\n")
   }
   
-  kkkk
-  cat("\n<<< building days...")
+  
+  if (verbose) cat("\n<<< building days...")
   if (length(palmsplus) > 0 & length(palmsplus_domains) > 0 & length(palmsplus_fields) &
-      length(home) > 0 & length(school) > 0 & length(home_nbh) > 0 & length(school_nbh) > 0 &
-      length(participant_basis) > 0) {
+      all(Nlocation_objects > 0) & length(participant_basis) > 0) {
     days <- hbt_build_days(data = palmsplus,
                            palmsplus_domains = palmsplus_domains,
                            palmsplus_fields = palmsplus_fields,
                            loca = loca,
-                           participant_basis = participant_basis)
+                           participant_basis = participant_basis,
+                           verbose = verbose)
+    if (length(days) > 0) {
+      if (verbose) cat(paste0("  N rows in days object: ", nrow(days)))
+    } else {
+      if (verbose) cat(paste0("  WARNING: no days object produced."))
+    }
     write_csv(days,  file = fns[2])
     # sf::st_write(palmsplus, dsn = paste0(palmsplus_folder, "/", dataset_name, "_palmsplus.shp"), append = FALSE)
-    cat("done>>>\n")
+    if (verbose) cat(">>>\n")
   } else {
-    cat("skipped because insufficient input data>>>\n")
+    if (verbose) cat("skipped because insufficient input data>>>\n")
   }
   
-  
   trajectory_locations = trajectory_locations[order(trajectory_locations$name),]
-  cat("\n<<< building trajectories...")
+  if (verbose) cat("\n<<< building trajectories...\n")
   if (length(palmsplus) > 0 & length(trajectory_fields) > 0) {
+    
     trajectories <- hbt_build_trajectories(data = palmsplus,
                                            trajectory_fields = trajectory_fields,
                                            trajectory_locations = trajectory_locations)
-    
-    write_csv(trajectories,  file = fns[3])
-    shp_file = paste0(palmsplus_folder, "/", dataset_name, "_trajecories.shp")
-    if (file.exists(shp_file)) file.remove(shp_file) # remove because st_write does not know how to overwrite
-    sf::st_write(obj = trajectories, dsn = shp_file)
-    cat("done>>>\n")
+    if (length(trajectories) > 0) {
+      write_csv(trajectories,  file = fns[3])
+      shp_file = paste0(palmsplus_folder, "/", dataset_name, "_trajecories.shp")
+      if (file.exists(shp_file)) file.remove(shp_file) # remove because st_write does not know how to overwrite
+      
+      sf::st_write(obj = trajectories, dsn = shp_file)
+      if (verbose) cat(paste0("  N rows in trajectories object: ", nrow(trajectories)))
+    } else {
+      if (verbose) cat(paste0("  WARNING: no trajectories object produced."))
+    }
+    if (verbose) cat(">>>\n")
   } else {
-    cat("skipped because insufficient input data>>>\n")
+    if (verbose) cat("skipped because insufficient input data>>>\n")
   }
-  cat("\n<<< building multimodal...")
+  if (verbose) cat("\n<<< building multimodal...\n")
   if (length(palmsplus) > 0 & length(multimodal_fields) > 0 & length(trajectory_locations) > 0) {
     multimodal <- hbt_build_multimodal(data = trajectories,
                                        spatial_threshold = 200,
                                        temporal_threshold = 10,
                                        palmsplus = palmsplus,
                                        multimodal_fields = multimodal_fields,
-                                       trajectory_locations = trajectory_locations)
+                                       trajectory_locations = trajectory_locations,
+                                       verbose = verbose)
+    
     if (length(multimodal) > 0) {
       write_csv(multimodal, file = fns[4])
       shp_file = paste0(palmsplus_folder, "/", dataset_name, "_multimodal.shp")
       if (file.exists(shp_file)) file.remove(shp_file) # remove because st_write does not know how to overwrite
       sf::st_write(obj = multimodal, dsn = shp_file)
+      if (verbose) cat(paste0("  N rows in multimodal object: ", nrow(multimodal)))
+    } else {
+      if (verbose) cat(paste0("  WARNING: no multimodal object produced."))
     }
-    cat("done>>>\n")
+    if (verbose) cat(">>>\n")
   } else {
-    cat("skipped because insufficient input data>>>\n")
+    if (verbose) cat("skipped because insufficient input data>>>\n")
   }
   
   return()
